mcp_servers/requirements.txt
mcp[cli]>=1.2.0
uvicorn>=0.30.0
google-api-python-client>=2.150.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
